clean up part 2
